requirements.txt
pytest
selenium
requests
psycopg2-binary
pytest-html
pytest-xdist
responses